changed median to return a double instead of an int

+++ Q1/question1.c
@@ -4,5 +4,5 @@
 #include <math.h> 
 
 double mean(int* x, int size); 
-int median(int* x, int size); 
+double median(int* x, int size); 
 int mode(int* x, int size);
